backend/requirements.txt
# =====================================
# ТОВАРИЩБОТ - СОВМЕСТИМЫЕ ЗАВИСИМОСТИ
# =====================================
# Проверенные версии без конфликтов зависимостей

# =====================================
# ОСНОВНЫЕ ЗАВИСИМОСТИ FASTAPI
# =====================================
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
aiofiles==23.2.1
pydantic==2.5.0
typing-extensions==4.8.0

# HTTP и сеть (совместимые с FastAPI)
anyio>=3.7.1,<4.0.0  # Совместимо с FastAPI 0.104.1
sniffio==1.3.0
idna==3.6
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
h11==0.14.0
starlette==0.27.0
httpx==0.25.2

# =====================================
# OPENAI API
# =====================================
openai==1.54.3
aiohttp==3.9.1

# =====================================
# MYSQL DATABASE INTEGRATION
# =====================================
aiomysql==0.2.0
PyMySQL==1.1.0
SQLAlchemy==2.0.23
alembic==1.13.1

# =====================================
# БЕЗОПАСНОСТЬ И АВТОРИЗАЦИЯ
# =====================================
PyJWT==2.8.0
cryptography>=42.0.0,<46.0.0  # Доступная версия
bcrypt==4.1.2
passlib[bcrypt]==1.7.4

# =====================================
# ВАЛИДАЦИЯ ДАННЫХ
# =====================================
validators==0.22.0
email-validator==2.1.0

# =====================================
# ОБРАБОТКА ФАЙЛОВ И ДОКУМЕНТОВ
# =====================================
Pillow==10.1.0
pillow-heif
python-magic-bin==0.4.14
PyPDF2==3.0.1
PyMuPDF==1.23.14
python-docx==1.1.0
pandas==2.1.4
openpyxl==3.1.2
xlrd==2.0.1
striprtf

# =====================================
# TELEGRAM INTEGRATION
# =====================================
python-telegram-bot==20.7
aiogram==3.2.0

# =====================================
# ЛОГИРОВАНИЕ И МОНИТОРИНГ
# =====================================
structlog==23.2.0
colorlog==6.8.0
psutil==5.9.6

# =====================================
# JSON И СЕРИАЛИЗАЦИЯ
# =====================================
orjson==3.9.10
ujson==5.9.0

# =====================================
# РАБОТА С ДАТАМИ
# =====================================
python-dateutil==2.8.2
pytz==2023.3

# =====================================
# ДОПОЛНИТЕЛЬНЫЕ УТИЛИТЫ
# =====================================
pydantic-settings==2.1.0
python-decouple==3.8
rich==13.7.0
tenacity==8.2.3

PyJWT==2.8.0

fastapi-csrf-protect==0.3.4
requests

APScheduler